+++ Extensions/ServiceExtensions.cs
@@ -1,3 +1,5 @@
+using System.ComponentModel;
+using System.Globalization;
 using System.Reflection;
 using Autofac;
 using Autofac.Extensions.DependencyInjection;
@@ -38,8 +40,8 @@ public static class ServiceExtensions
     {
         // Register core services directly for non-Autofac scenarios
         services.AddScoped<IViewPageRenderer, ViewPageRenderer>();
-        services.AddScoped<IIdevsPdfExporter, IdevsPdfExporter>();
         services.AddScoped<IIdevsExcelExporter, IdevsExcelExporter>();
+        services.AddScoped<IIdevsPdfExporter>(_ => new IdevsPdfExporter(CultureInfo.CurrentCulture));
 
         // Register attribute-based services
         RegisterAttributeBasedServices(services);